Troca do tipo INT para FLOAT

+++ include/Distancias.h
@@ -1,1 +1,1 @@
-void distEuclid(int p[], int q[], int k, int *distE);
+void distEuclid(float p[], float q[], int k, float *distE);
